move bot forms components
