+++ packages/frontend/src/components/CreateBot/index.tsx
@@ -1,9 +1,9 @@
 import React, {FunctionComponent} from 'react';
 import WrappedSelect from "@/components/BaseFormFields/Select/WrappedSelect";
 import SelectOption from "@/types/SelectOption";
-import RetrievalConversational from "@/components/CreateBot/RetrievalConversational";
+import RetrievalConversational from "../BotForms/RetrievalConversational";
 import {AGENT_OPTIONS} from "@/constants";
-import Conversational from "@/components/CreateBot/Conversational";
+import Conversational from "../BotForms/Conversational";
 import styles from './styles.module.scss';
 
 const CreateBot: FunctionComponent = () => {
